refactor: server 디렉토리명을 serverState로 수정
