+++ src/App.tsx
@@ -1,6 +1,6 @@
 import { QueryClientProvider } from "@tanstack/react-query";
 import ProductFeedPage from "@pages/ProductFeedPage";
-import { queryClient } from "@server/queryClient";
+import { queryClient } from "@serverState/queryClient";
 
 function App() {
   return (
