+++ src/serverState/__test__/useCreateCartItemMutation.test.tsx
@@ -1,10 +1,10 @@
 import { ReactNode, act } from "react";
 import { vi } from "vitest";
 import { QueryClientProvider } from "@tanstack/react-query";
-import { useUpdateCartItemMutation } from "@src/server/mutations/useUpdateCartItemMutation";
+import { useCreateCartItemMutation } from "@serverState/mutations/useCreateCartItemMutation";
 import { renderHook, waitFor } from "@testing-library/react";
-import { QUERY_KEYS } from "@server/__constants__/queryKeys";
-import { queryClient } from "@server/queryClient";
+import { QUERY_KEYS } from "@serverState/__constants__/queryKeys";
+import { queryClient } from "@serverState/queryClient";
 import { server } from "@src/mocks/server";
 import { HttpResponse, http } from "msw";
 import { CART_API_URL } from "@src/env/envVariables";
@@ -14,19 +14,19 @@ const wrapper = ({ children }: { children: ReactNode }) => (
   <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
 );
 
-describe("useUpdateCartItemMutation", () => {
+describe("useCreateCartItemMutation", () => {
   beforeEach(() => {
     vi.clearAllMocks();
   });
 
-  it("cartItem 수정 성공 시 CartItems 쿼리 무효화를 수행한다.", async () => {
+  it("cartItem 생성 성공 시 CartItems 쿼리 무효화를 수행한다.", async () => {
     const invalidateQueriesSpy = vi.spyOn(queryClient, "invalidateQueries");
 
-    const { result } = renderHook(() => useUpdateCartItemMutation(), {
+    const { result } = renderHook(() => useCreateCartItemMutation(), {
       wrapper,
     });
 
-    const params = { cartItemId: 101, quantity: 1 };
+    const params = { productId: 101, quantity: 1 };
 
     await act(async () => {
       result.current.mutate(params);
@@ -37,20 +37,20 @@ describe("useUpdateCartItemMutation", () => {
     });
   });
 
-  it("cartItem 수정 실패 시 에러 핸들링을 수행한다.", async () => {
+  it("cartItem 생성 실패 시 에러 핸들링을 수행한다.", async () => {
     server.use(
-      http.patch(`${CART_API_URL}${API_URL.cartItems}/:cartItemId`, () => {
+      http.post(CART_API_URL + API_URL.cartItems, () => {
         return new HttpResponse(null, { status: 500 });
       })
     );
 
     const onError = vi.fn();
 
-    const { result } = renderHook(() => useUpdateCartItemMutation(onError), {
+    const { result } = renderHook(() => useCreateCartItemMutation(onError), {
       wrapper,
     });
 
-    const params = { cartItemId: 101, quantity: 1 };
+    const params = { productId: 101, quantity: 1 };
 
     await act(async () => {
       result.current.mutate(params);
