+++ src/serverState/__test__/useDeleteCartItemMutation.test.tsx
@@ -1,35 +1,35 @@
 import { ReactNode, act } from "react";
 import { vi } from "vitest";
 import { QueryClientProvider } from "@tanstack/react-query";
-import { useCreateCartItemMutation } from "@src/server/mutations/useCreateCartItemMutation";
 import { renderHook, waitFor } from "@testing-library/react";
-import { QUERY_KEYS } from "@server/__constants__/queryKeys";
-import { queryClient } from "@server/queryClient";
-import { server } from "@src/mocks/server";
+import { QUERY_KEYS } from "@serverState/__constants__/queryKeys";
+import { queryClient } from "@serverState/queryClient";
+import { server } from "@mocks/server";
 import { HttpResponse, http } from "msw";
-import { CART_API_URL } from "@src/env/envVariables";
-import { API_URL } from "@src/apis/__constants__/apiUrl";
+import { CART_API_URL } from "@env/envVariables";
+import { API_URL } from "@apis/__constants__/apiUrl";
+import { useDeleteCartItemMutation } from "@serverState/mutations/useDeleteCartItemMutation";
 
 const wrapper = ({ children }: { children: ReactNode }) => (
   <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
 );
 
-describe("useCreateCartItemMutation", () => {
+describe("useDeleteCartItemMutation", () => {
   beforeEach(() => {
     vi.clearAllMocks();
   });
 
-  it("cartItem 생성 성공 시 CartItems 쿼리 무효화를 수행한다.", async () => {
+  it("cartItem 삭제 성공 시 CartItems 쿼리 무효화를 수행한다.", async () => {
     const invalidateQueriesSpy = vi.spyOn(queryClient, "invalidateQueries");
 
-    const { result } = renderHook(() => useCreateCartItemMutation(), {
+    const { result } = renderHook(() => useDeleteCartItemMutation(), {
       wrapper,
     });
 
-    const params = { productId: 101, quantity: 1 };
+    const ITEM_ID = 101;
 
     await act(async () => {
-      result.current.mutate(params);
+      result.current.mutate(ITEM_ID);
     });
 
     await waitFor(() => {
@@ -37,23 +37,23 @@ describe("useCreateCartItemMutation", () => {
     });
   });
 
-  it("cartItem 생성 실패 시 에러 핸들링을 수행한다.", async () => {
+  it("cartItem 삭제 실패 시 에러 핸들링을 수행한다.", async () => {
     server.use(
-      http.post(CART_API_URL + API_URL.cartItems, () => {
+      http.delete(`${CART_API_URL}${API_URL.cartItems}/:cartItemId`, () => {
         return new HttpResponse(null, { status: 500 });
       })
     );
 
     const onError = vi.fn();
 
-    const { result } = renderHook(() => useCreateCartItemMutation(onError), {
+    const { result } = renderHook(() => useDeleteCartItemMutation(onError), {
       wrapper,
     });
 
-    const params = { productId: 101, quantity: 1 };
+    const ITEM_ID = 101;
 
     await act(async () => {
-      result.current.mutate(params);
+      result.current.mutate(ITEM_ID);
     });
 
     await waitFor(() => {
