+++ src/serverState/mutations/useCreateCartItemMutation.ts
@@ -1,7 +1,7 @@
 import { CreateCartItemParams, createCartItem } from "@src/apis/cartItems";
 import { useMutation } from "@tanstack/react-query";
-import { MUTATION_KEYS } from "@server/__constants__/queryKeys";
-import { queryInvalidator } from "@server/queryClient";
+import { MUTATION_KEYS } from "@serverState/__constants__/queryKeys";
+import { queryInvalidator } from "@serverState/queryClient";
 import type { OnError } from "onError";
 
 export const useCreateCartItemMutation = (onError?: OnError) => {
